move cli entry points to <module>/__main__.py
