--- cs251tk/toolkit/__init__.py
@@ -1,9 +1,0 @@
-from .args import process_args
-from .config import conf
-from .find_update import update_available
-from .get_students import get_students
-from .gist import post_gist
-from .progress_bar import progress_bar
-from .process_student import process_student
-from .save_recordings import save_recordings, gist_recordings
-from .tabulate import tabulate, asciiify
